Crud Operation and Exception Handling

+++ product-service/src/main/java/com/ecomapp/productservice/service/ProductServiceImpl.java
@@ -1,5 +1,7 @@
 package com.ecomapp.productservice.service;
 
+import com.ecomapp.productservice.exception.DuplicateItemException;
+import com.ecomapp.productservice.exception.ItemNotFoundException;
 import com.ecomapp.productservice.model.Product;
 import com.ecomapp.productservice.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +19,10 @@ public class ProductServiceImpl implements ProductService {
     }
 
     public Product addProduct(Product product) {
+
+        if (productRepo.existsById(product.getId())){
+            throw new DuplicateItemException("Product with id "+product.getId()+ " Already Present");
+        }
         return productRepo.save(product);
     }
 
@@ -25,12 +31,19 @@ public class ProductServiceImpl implements ProductService {
     }
 
     public Product getProductById(long id) {
-        return productRepo.findById(id).get();
+        return productRepo.findById(id).orElseThrow(() -> new ItemNotFoundException("Item with id "+id+ " Not found"));
     }
 
     public void deleteProduct(long id) {
 
+        if (!productRepo.existsById(id)){
+            throw new ItemNotFoundException("Item with id "+id+ " Not found");
+        }
         productRepo.deleteById(id);
 
+    }
+
+    public List<Product> searchByName(String name) {
+        return productRepo.findByNameContainingIgnoreCase(name);
     }
 }
